fix(build): fix assets url generation during build process
In moving to the new web pipeline to publish the component assets, we adopted the new cached asset
mechanism which uses applicaiton-relative paths to fetch assets. Since the components are not an
app, but are included in other apps, these paths were not correct. This changes the build so it will
use a fully qualified URL to look up assets.

COMUI-1063